+++ NEWS.md
@@ -1,5 +1,6 @@
 # big.data.table 0.3.3 (devel)
 
+* logging of `in_rows` in nodes and client included in logR call.
 * processing of `outer.aggregate` push down to `bdt.eval` to include in timing of client task.
 * `outer.aggregate` arg can accept function.
 * new function `bdt.eval.log` just for handling logR and nicely parsed expressions to logs.